+++ src/test/java/uk/gov/service/notify/JsonUtilsTest.java
@@ -65,16 +65,14 @@ public class JsonUtilsTest {
     }
 
     @Test
-    public void testJsonToMap_ifNull() throws Exception {
-        JSONObject json = null;
-
-        Map map = JsonUtils.jsonToMap(json);
+    public void testJsonToMap_ifNull() {
+        Map map = JsonUtils.jsonToMap(null);
 
         assertEquals(new HashMap<>(), map);
     }
 
     @Test
-    public void testJsonToMap_ifJSONNull() throws Exception {
+    public void testJsonToMap_ifJSONNull() {
         JSONObject json = new JSONObject(JSONObject.NULL);
 
         Map map = JsonUtils.jsonToMap(json);
@@ -83,7 +81,7 @@ public class JsonUtilsTest {
     }
 
     @Test
-    public void testJsonToMap_ifMapNull() throws Exception {
+    public void testJsonToMap_ifMapNull() {
         JSONObject json = new JSONObject((Map)null);
 
         Map map = JsonUtils.jsonToMap(json);
@@ -92,7 +90,7 @@ public class JsonUtilsTest {
     }
 
     @Test
-    public void testJsonToMap_fromJsonObject() throws Exception {
+    public void testJsonToMap_fromJsonObject() {
         JSONObject json = new JSONObject();
         json.put("aKey", "aValue");
         json.put("bKey", new JSONObject("{\"bbKey\": \"bbValue\"}"));
@@ -105,7 +103,7 @@ public class JsonUtilsTest {
    }
 
     @Test
-    public void testJsonToMap_fromJsonObjectFromMap() throws Exception {
+    public void testJsonToMap_fromJsonObjectFromMap() {
 
         JSONObject json = new JSONObject(expectedMap);
 
@@ -115,7 +113,7 @@ public class JsonUtilsTest {
     }
 
     @Test
-    public void testJsonToMap_fromString_ifEmpty() throws Exception {
+    public void testJsonToMap_fromString_ifEmpty() {
         JSONObject json = new JSONObject("{}");
 
         Map map = JsonUtils.jsonToMap(json);
@@ -124,7 +122,7 @@ public class JsonUtilsTest {
     }
 
     @Test
-    public void testJsonToMap_fromString() throws Exception {
+    public void testJsonToMap_fromString() {
         JSONObject json = new JSONObject("{\n" +
                 "\"aKey\": \"aValue\",\n" +
                 "\"bKey\": {\"bbKey\": \"bbValue\"},\n" +
@@ -138,7 +136,7 @@ public class JsonUtilsTest {
     }
 
     @Test
-    public void testJsonToMap_orderDoesNotMatter() throws Exception {
+    public void testJsonToMap_orderDoesNotMatter() {
         JSONObject json = new JSONObject("{\n" +
                 "\"bKey\": {\"bbKey\": \"bbValue\"},\n" +
                 "\"aKey\": \"aValue\",\n" +
@@ -151,19 +149,15 @@ public class JsonUtilsTest {
         assertEquals(expectedMap, map);
     }
 
-
-
     @Test
-    public void testJsonToList_ifNull() throws Exception {
-        JSONArray json = null;
-
-        List list = JsonUtils.jsonToList(json);
+    public void testJsonToList_ifNull() {
+        List list = JsonUtils.jsonToList(null);
 
         assertEquals(new ArrayList<>(), list);
     }
 
     @Test
-    public void testJsonToList_ifListNull() throws Exception {
+    public void testJsonToList_ifListNull() {
         JSONArray json = new JSONArray((List)null);
 
         List list = JsonUtils.jsonToList(json);
@@ -172,7 +166,7 @@ public class JsonUtilsTest {
     }
 
     @Test
-    public void testJsonToList_fromJsonArray() throws Exception {
+    public void testJsonToList_fromJsonArray() {
         JSONArray json = new JSONArray();
         json.put("dValue");
         json.put(true);
@@ -187,7 +181,7 @@ public class JsonUtilsTest {
     }
 
     @Test
-    public void testJsonToList_fromJsonArrayFromList() throws Exception {
+    public void testJsonToList_fromJsonArrayFromList() {
         JSONArray json = new JSONArray(expectedList);
 
         List list = JsonUtils.jsonToList(json);
@@ -196,7 +190,7 @@ public class JsonUtilsTest {
     }
 
     @Test
-    public void testJsonToList_fromString_ifEmpty() throws Exception {
+    public void testJsonToList_fromString_ifEmpty() {
         JSONArray json = new JSONArray("[]");
 
         List list = JsonUtils.jsonToList(json);
@@ -205,7 +199,7 @@ public class JsonUtilsTest {
     }
 
     @Test
-    public void testJsonToList_fromString() throws Exception {
+    public void testJsonToList_fromString() {
         JSONArray json = new JSONArray("[\n" +
                 "\"dValue\", \n" +
                 "true, \n" +
@@ -226,7 +220,7 @@ public class JsonUtilsTest {
     }
 
     @Test
-    public void testJsonToList_orderMatters() throws Exception {
+    public void testJsonToList_orderMatters() {
         JSONArray json = new JSONArray("[\n" +
                 "\"dValue\", \n" +
                 "true, \n" +